feat: milliseconds, currentMilliseconds의 타입을 Int64로 변경

+++ Clock/Clock/Data/Model/Timer/TimerEntity.swift
@@ -10,8 +10,8 @@ import CoreData
 @objc(Timer)
 public class TimerEntity: NSManagedObject {
     @NSManaged public var id: UUID
-    @NSManaged public var milliseconds: Int16
-    @NSManaged public var currentMilliseonds: Int16
+    @NSManaged public var milliseconds: Int64
+    @NSManaged public var currentMilliseonds: Int64
     @NSManaged public var label: String
     @NSManaged public var sound: String
 }
